Ensure dbReads respect list order.
On dbReads for lists, make sure we sort by the ids written
with each entry.  This guarantees reads come back in the same
order the write was made, to handle the lack of ordering in a
JSON scheme.  This fix is for improper vm autostarting, but
is relevant for any toolstack dbRead using lists.

Signed-off-by: Chris Rogers <rogersc@ainfosec.com>